Rename solve_strategy
