+++ _deprecated_src/collect_execution_pickles.py
@@ -5,7 +5,7 @@ from glob import glob
 
 import pandas as pd
 
-from remat.core.solvers.strategy_enum import SolveStrategy
+from remat.core.solvers.enum_strategy import SolveStrategy
 
 
 def extract_params():
