+++ _deprecated_src/evaluation/util/evaluation_utils.py
@@ -7,7 +7,7 @@ import pandas
 import ray
 from tqdm import tqdm
 
-from remat.core.solvers.strategy_enum import SolveStrategy
+from remat.core.solvers.enum_strategy import SolveStrategy
 from solvers.result import RSResult
 
 RSResultDict = Dict[SolveStrategy, List[RSResult]]
