+++ _deprecated_src/execute_one.py
@@ -8,7 +8,7 @@ import pickle
 
 import dotenv
 
-from remat.core.solvers.strategy_enum import SolveStrategy
+from remat.core.solvers.enum_strategy import SolveStrategy
 from integration.tf2.extraction import MODEL_NAMES
 from evaluation.eval_execution import execute_one
 from utils.setup_logger import setup_logger
